Split PersonaPresence
