+++ packages/office-ui-fabric-react/src/components/Persona/Persona.types.ts
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { IRenderFunction } from '../../Utilities';
 import { PersonaBase } from './Persona.base';
 import { PersonaCoinBase } from './PersonaCoin.base';
-import { PersonaPresenceBase } from './PersonaPresence';
+import { PersonaPresenceBase } from './PersonaPresence.base';
 import { ImageLoadState } from '../../Image';
 import { IStyle, ITheme } from '../../Styling';
 import { IStyleFunction } from '../../Utilities';
